reorganise codebase a bit
