+++ packages/kit/src/exports/vite/types.d.ts
