+++ packages/kit/src/runtime/server/page/index.js
@@ -3,8 +3,8 @@ import { render_response } from './render.js';
 import { respond_with_error } from './respond_with_error.js';
 import { method_not_allowed, error_to_pojo, allowed_methods } from '../utils.js';
 import { create_fetch } from './fetch.js';
-import { HttpError, Redirect } from '../../../index/private.js';
-import { error, json } from '../../../index/index.js';
+import { HttpError, Redirect } from '../../control.js';
+import { error, json } from '../../../exports/index.js';
 import { compact } from '../../../utils/array.js';
 import { normalize_error } from '../../../utils/error.js';
 import { load_data, load_server_data } from './load_data.js';
